Fix link rot in partitioning page (#13)
"A beginner’s guide to disks and disk partitions in Linux" is put behind
a soft paywall. Replaced it with a different link to the same article.

"Understanding the Linux File System" is no longer available. Replaced
it the file-hierarchy manpage which covers the same information in a
similar way.